+++ src/dockerfiles/diegomura__react-pdf-1552.Dockerfile
@@ -52,6 +52,7 @@ ENV DBUS_SESSION_BUS_ADDRESS="unix:path=/run/dbus/system_bus_socket"
 RUN dbus-daemon --system --fork
 
 ENV PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true
+ENV PUPPETEER_SKIP_DOWNLOAD=true
 ENV OPENSSL_CONF /etc/ssl
 
 RUN useradd -m chromeuser
@@ -91,36 +92,32 @@ python2 -V
 EOF_a32284fbdcc9
 
 
-RUN <<EOF_0349781138e9
+RUN <<EOF_1ae8f2f2616a
 #!/bin/bash
 set -euxo pipefail
 git clone -o origin https://github.com/diegomura/react-pdf /testbed
-chmod -R 777 /testbed
 cd /testbed
 git reset --hard 7acd39fa1b60d2379e48584964d8c4643aa473be
 git remote remove origin
-TARGET_EPOCH=$(git show -s --format=%ct 7acd39fa1b60d2379e48584964d8c4643aa473be)
-git tag -l | while read tag; do TAG_COMMIT=$(git rev-list -n 1 "$tag"); TAG_EPOCH=$(git show -s --format=%ct "$TAG_COMMIT"); if [ "$TAG_EPOCH" -gt "$TARGET_EPOCH" ]; then git tag -d "$tag"; fi; done
-git branch -D $(git branch | grep -v "^\*") 2>/dev/null || true
+git branch | grep -v '^\*' | xargs -r git branch -D || true
+git tag -l | xargs -r git tag -d
 git reflog expire --expire=now --all
+git gc --prune=now --aggressive
+TARGET_EPOCH=$(git show -s --format=%ct 7acd39fa1b60d2379e48584964d8c4643aa473be)
+AFTER_EPOCH=$((TARGET_EPOCH + 1))
+AFTER_TIMESTAMP=$(date -u -d "@$AFTER_EPOCH" "+%Y-%m-%d %H:%M:%S")
+COMMIT_COUNT=$(git log --oneline --all --since="$AFTER_TIMESTAMP" | wc -l)
+[ "$COMMIT_COUNT" -eq 0 ] || exit 1
 cd - || true
+chmod -R 777 /testbed
 cd /testbed
 git clean -fdxq
 source $NVM_DIR/nvm.sh
 npm i -g yarn
 yarn install
-EOF_0349781138e9
+EOF_1ae8f2f2616a
 
 
-RUN <<EOF_ec97711334a2
-#!/bin/bash
-set -euxo pipefail
-mkdir -p /swebench/image_assets
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/132485670-9c9653b2-e563-438d-b73b-461cc16ed88b.png' 'https://user-images.githubusercontent.com/1511512/132485670-9c9653b2-e563-438d-b73b-461cc16ed88b.png' || true
-mkdir -p /swebench/image_assets/problem_statement
-curl -fsSL -o '/swebench/image_assets/problem_statement/132486410-efc68c57-bc52-4ce3-80b5-c2a67a3e058d.png' 'https://user-images.githubusercontent.com/1511512/132486410-efc68c57-bc52-4ce3-80b5-c2a67a3e058d.png' || true
-EOF_ec97711334a2
-
+COPY src/image_assets/diegomura__react-pdf-1552/ /swebench/image_assets/
 
 WORKDIR /testbed
